+++ CHANGES.rst
@@ -1,5 +1,7 @@
 v0.4.0
 ======
+- OdeSys.integrate_* made private.
+- symbolic.PartiallySolvedSystem added
 - multiple (chained) pre and postprocessors supported
 - stiffness may be inspected retroactively (ratio biggest/smallest eigenvalue 
   of the jacobian matrix).